# Conflicts:
#	src/main/java/at/hannibal2/skyhanni/utils/renderables/Renderable.kt

+++ src/main/java/at/hannibal2/skyhanni/utils/tracker/SkyHanniItemTracker.kt
@@ -178,7 +178,7 @@ class SkyHanniItemTracker<Data : ItemTrackerData>(
     }
 
     fun addTotalProfit(profit: Double, totalAmount: Long, action: String): Renderable {
-        val profitFormat = profit.addSeparators()
+        val profitFormat = profit.toInt().addSeparators()
         val profitPrefix = if (profit < 0) "§c" else "§6"
 
         val tips = if (totalAmount > 0) {
